wrote code to get the latest itinerary for the home page
